+++ src/public/assets/talk-WSEiVtNq.js
@@ -1,4 +1,4 @@
-import{c as S,r,j as e,C as V,h as Y,m as le,L as Z,M as U,a as A,n as de,T as G,o as Q,H as he,p as ue,F as me,B as C,q as Me,s as Ce}from"./card-CswwyPQ7.js";/**
+import{c as S,r,j as e,C as V,h as Y,m as le,L as Z,M as U,a as A,n as de,T as G,o as Q,H as he,p as ue,F as me,B as C,q as Me,s as Ce}from"./card-Cxx4UyZ5.js";/**
  * @license lucide-react v0.462.0 - ISC
  *
  * This source code is licensed under the ISC license.